requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
openai==0.28.1
python-dotenv==1.0.0